// modules/client/utilities/DataTypes.ts
export interface ProcessDataExport {
  MetaDataId: string;
  Asset: string;
  IdentifierCode: string;
  IdentifierCode2: string;
  PartNumber: string;
  OpEndTime: Date;
  PassFail: boolean;
  OperationId: string;
  Line: string;
  Label: string;
  Operator: string;
  Description: string;
  CycleTime: string;
  Revision: string;
  Sender: string;
  TestPlan: string;
  Barcode: string;
}

export interface ProcessDataOperator {
  id: number;
  Asset: string;
  PartNumber: string;
  Date: Date;
  StartTime: Date;
  EndTime: Date;
  Passes: number;
  Fails: number;
  OperationId: string;
  Line: string;
  Label: string;
  Operator: string;
  Description: string;
  CycleTime: string;
  Revision: string;
  Sender: string;
  TestPlan: string;
}

export interface ProcessDataOperatorTotals {
  id: number;
  Asset: string;
  PartNumber: string;
  Date: Date;
  StartTime: Date;
  EndTime: Date;
  Passes: number;
  Fails: number;
  OperationId: string;
  Line: string;
  Label: string;
  Operator: string;
  Revision: string;
  Sender: string;
  TestPlan: string;
  CycleTime: number;
  RunActual: number;
  RunTheory: number;
  Efficiency: number;
  PartsPerHour: number;
}